Move `getenv_use_multiple_threads` to `qiskit-util`
This is a trivially shared utility function with nothing to do with the
core IR, and can easily move into the base utility crate.
